Moved the creators menu into a separate package
